Updated openstack/openstack Project: openstack/nova  840136a9411ca49e25e83f4aa6007e0cc1b67dca
Update HostManager and filters to use ComputeNode ratios

Since the ComputeNode object provides a compatible facade for the Scheduler,
we can get it from the HostManager and provide those fields up to the filters.

As the HostManager is calling ComputeNodeList.get_all(), then the related HostStates
will all have the scheduler nova.conf allocation ratios.

Change-Id: I3bd28cd2069ada2f9b0d1fd9c05d12bb6f8f75d9
Partially-Implements: blueprint allocation-ratio-to-resource-tracker