Updated openstack/openstack Project: openstack/nova  28747d15b0cbfe91bc3ec8f86b0f59c3c1936938
Update ComputeNode values with allocation ratios in the RT

Now that we have added the fields for persisting the allocation ratios, we can
have the ResouceTracker persisting them by adding them to the local ComputeNode
object which is persisted by calling the _update() method.
It will then send by default 0.0 unless the operator explicitely specified an
allocation ratio in the compute nova.conf.

Thanks to the ComputeNode object hydratation on the scheduler side, the facade
will make sure that if a default 0.0 is provided by either a compute node or
by the scheduler's nova.conf, it will actually get the original allocation
ratios (ie. 16.0 and 1.5)
Since the Scheduler reads the same RT opts but goes thru the ComputeNode object,
it will also get the Facade returning 16.0 and 1.5 unless the operator
explicitely provided other ratios for the scheduler's nova.conf

That compat mode (having ratios defaulted to 0.0) is only planned to be kept for
Liberty and will be removed in the next release (Mitaka) in order to leave to
the operators a whole cycle to update all their computes' nova.conf files
if they want to have a specific allocation ratio for their compute nodes.

Note: that BP now allows to provide a specific allocation ratio per compute.

DocImpact
UpgradeImpact

Change-Id: I382421968d4c7989a6ff08c74c21139bf6be12e5
Implements: blueprint allocation-ratio-to-resource-tracker