Updated openstack/openstack Project: openstack/openstack-doc-tools  92dda518f4407aaf9aad28c3cf80e0f9081aad43
openstack-auto-commands: fix issues with python-glanceclient 1.0.0

* explicitly set --os-image-api-version 1

This is necessary because python-glanceclient 1.0.0 will use v2 when v2
is accessible.

* only extract subcommands from lastline of bash-completion output

This is necessary because 'glance bash-completion' outputs a
warning starting with python-glanceclient 1.0.0.

---snip---
WARNING: The client is falling back to v1 because the accessing to v2
failed. This behavior will be removed in future versions
---snap---

Change-Id: I936f5704e3badd4cb12d3d9c0753eb154bf594b2