Updated openstack/openstack Project: openstack-infra/project-config  e1efe6feb7a4b336b8bd4259557702c9d9a33279
Tripleo CI reshuffle

Tripleo is in the process of changing from tripleo-incubator/devtest
driven CI to CI driven by instack.

Until we are confident that the instack based CI is reliable and
gives us the coverage we want. We are removing our CI from all non tripleo
projects, with the exceptions of heat, ironic, neutron and nova as we have
found a lot of value in running running our CI on these projects in the past.
For these 4 projects the instack jobs will be available in the
experimental-tripleo queue.

Remove the ubuntu based jobs as instack doesn't currently support them
and there isn't currently any users interested in adding support.

Change-Id: I7cd8bd4f1aaba727b9f42c093f297c6688c92025