Updated openstack/openstack Project: openstack/keystone  222b02b37f85c27576c512b0aaf42e301adb29ac
Unit tests for is_domain field in project's table

This patch provides the unit tests for the creation of projects with
the is_domain=True field in the project's table. Since we don't have the
implementation yet, most of the tests are marked as WIP.
The actual implementation is in a follow-on patch.

Co-Authored-By: Henrique Truta <henrique@lsd.ufcg.edu.br>

Change-Id: Ic3c589103f1fbaf8193bc1dbe847ef57e544374a
Partially-Implements: bp reseller