Updated openstack/openstack Project: openstack/cinder  f517519d36e97f1cb0c9894d6e0d54a9dcc4165e
force_detach terminate_connection needs connector

The function force_detach in cinder/api/contrib/admin_actions does not
take a connector in the body of the request. When it calls terminate
connection it passes in an empty dict as the connector:
self.volume_api.terminate_connection(context, volume, {}, force=True)
This patch pulls the connector out of the body of the request, and
throws an exception if it is missing.

Change-Id: I14990eb4f42283231e4249a81d0b7fa18a8dfeae
Closes-Bug:#1485766