Updated openstack/openstack Project: openstack/openstack-manuals  1adaf4d147ef3e4f5b094333d8a76baa3529c684
Fix the procedure in Cloud Administrator Guide

After the procedure titled "Integrating Identity with multiple back
ends", there is a Note. After the Note, a new procedure appears to start
with "1. Define the destination LDAP server...". It is part of the
previous procedure, then the Note should be correctly indented so the
numbering continues correctly.

Change-Id: If10ba714f6a86124dc3de419dae017f4cf733c49
Closes-Bug: #1490791