Updated openstack/openstack Project: openstack/manila  14c18049f704b17f8a9b9579d5450d4202e2b957
Fix AttributeError without share type provided

A share_type is set to None, when creating a new share
without specifying a share type and a default_share_type
option is set to None.
Make scheduler check presense of a share type and
raise error if it is absent.
The raised error is logged finally.

Closes-Bug: #1477789
Change-Id: I6cdf3ce3940af4095b8e6afca4861ad1c2fc4795