Updated openstack/openstack Project: openstack-infra/puppet-gerrit  414b66a740e98726e4afcdca4340bc1accf6971a
Trigger reindex only when secondary index is set.

Gerrit only allows the reindex command to work when there is a
secondary index type configured. By default, on this module, the value
for the secondary index is `false`.

If we execute Gerrit's reindex command while the secondary index is
false it will cause an error on `puppet apply`.

This change extracts the reindex command into it's own step on the
process. The command will only execute if we configured it with a
secondary index.

Change-Id: Ifac5daec7db6c2452a731f03c69e52afd196b8da
Co-Authored-By: Bruno Tavares <btavare@thoughtworks.com>