Updated openstack/openstack Project: openstack/os-cloud-config  f08fea68805564f22f3a2ba0372820bceed87559
register-nodes: add support for fake_pxe

The fake_pxe driver in Ironic is a driver with a non-op power interface,
so all the power actions needs to be performed manually.

This updates the register-nodes command to support use of the fake_pxe
driver with Ironic.

The flake8 C901 was also added to the tox.ini to be ignored. This error
is raised because we have too much if..elif conditions as part of the
_extract_driver_info() function. It will need some refactoring later,
so ignoring the error for now.

Change-Id: I2de9643ddbdb73c5b49b7572cf476016d4911550