Updated openstack/openstack Project: openstack/nova  3feb88b2466cb36f382a99d76ed9f40d0519a74c
compute: improve exceptions related to disk size checks

Having read several user reports of FlavorDiskTooSmall, it is apparent
that this error is not meaningful to users. We create 3 new subclasses
of FlavorDiskTooSmall: FlavorDiskSmallerThanImage,
FlavorDiskSmallerThanMinDisk, and VolumeSmallerThanMinDisk, because
these exceptions require different actions by the user. The new
subclasses get a more detailed error message to help the user.

Change-Id: I2234c4f4f9b5ac0780b3ec2e6a168380c07ed8c1