Updated openstack/openstack Project: openstack/nova  c310255cb910ad08c7e67b7d8da8ba2fee5b54ef
Improve error logs for start/stop of locked instance

If someone tries to execute a command on a locked instance, we want them to
get the more explicit "Instance <uuid> is locked" message instead of the
more generic "Instance <uuid> is in an invalid state for <action>" message.

exception.InstanceIsLocked is a subclass of exception.InstanceInvalidState,
so we need to check for the more specific case first.

Change-Id: I33553e1a534a9aa1394a3927d745196476067c9d