Updated openstack/openstack Project: openstack/governance  8e577c50a17fb1ae5b82f322f1fb8a33df302b85
OpenStack Community App Catalog application to join the Big Tent

This is a request to add the App Catalog to the OpenStack
namespace.  The Community App Catalog is a collection of
ready-to-use applications, Heat templates and machine images, and
other OpenStack components that you can immediately deploy into
your OpenStack cloud.

The project was  developed in coordination with the OpenStack
Foundation, and is hosted by OpenStack Infra at
http://apps.openstack.org.  This project is already using all of
the OpenStack infrastructure and services and is gated through
gerrit.

We are using the both the Dev and Operators mailing list, and
bring up initiatives on the mailing lists in order to provide
ample opportunity for those in the community to provide feedback,
and so everyone can be kept aware of our progress and intentions.
We also hold weekly IRC meetings[1] which are announced
regularly, and logged for review.

The IRC channel #openstack-app-catalog is logged[2], and active
with community members, core developers and like minded people
all helping and discussing issues within the project as well as
questions that come up from using the App Catalog or adding
content to it.

Christopher Aedo has been acting as the temporary PTL, and has
been actively working with the App Catalog since before it was
officially launched at the Vancouver summit. We plan to hold an
official PTL election before the Tokyo summit.

This commit is a formal request to add the Community App Catalog
into the OpenStack namespace as an official project.

[1] http://eavesdrop.openstack.org/meetings/app_catalog/
[2] http://eavesdrop.openstack.org/irclogs/%23openstack-app-catalog/

Change-Id: I6092d2b574b4247da6328c96ae4d7f2e6a9c1941