Updated openstack/openstack Project: openstack/barbican  9e11a34fc19d721ca393191e9ee1af904f4a3e08
Ensure Project and ProjectQuotas Tables are in sync

Make some changes to clarify when the code is using the internal
project ID or the external project ID.

The SQL command ForeignKey marks a requirement that two tables are kept
in sync with a common key.  We want to ensure that for every project
quota record, there is a matching project record in the projects
table.

Also add the migration script that adds the project quotas table
to databases based on an older Barbican database model.

Change-Id: I860c87182d94967a290534e2825404f2df2e4033
Implements: blueprint quota-support-on-barbican-resources