Updated openstack/openstack Project: openstack/ceilometer  40be122a8b73c231af62811f0d74fc7c6268ea37
Event filtering for non-admin users

Non-admin users should be restricted to see only those
events that were generated by them and not the events
associated with other users.The current behavior doesn't
have any restrictions on the events that a non-admin user
can view.As part of this patch, a non-admin user will see
only the events that are associated with the user_id and
project_id that the user requesting events belongs to.
Events stored in the ceilometer database has information
from all other services and in a cloud environment, there
must be restrictions on the data a non-admin user can access.

Change-Id: I2ed7425c14a70d1db4621a277ba5f4f575a15eba
blueprint: events-rbac