Updated openstack/openstack Project: openstack/python-keystoneclient  19258abab202ee3d06eaf4797ed27bdc334a2c2d
Proper deprecation for httpclient.USER_AGENT

httpclient's USER_AGENT constant wasn't properly deprecated since
the deprecation was only a comment. Proper deprecation requires use
of warnings/debtcollector and documentation, but since this is a
module symbol there's no way to put out a warning, so it's just
documentation in this case.

bp deprecations

Change-Id: I02c77c690a31eea40935d2756748382abec86867