Updated openstack/openstack Project: openstack/barbican  98ad34d354d39e769591ec06536536620bdd12e0
Add Request ID to worker-side logs in barbican.

This can help correlate errors in logs to an incoming request
The request ID is additionally added to the RPC task args that is put into
the queue by the server. The client worker then includes the request ID
in the logs when it performs the task.

Change-Id: Ic359228cd7877022530be21c7d00996d5723a8e0
Closes-bug: #1459800