Updated openstack/openstack Project: openstack/heat  97572d146dff4d404abe0923d1f6cd88ccd31c80
enabling the stevedore.sphinxext extension.

The point behind the addition of the stevedore.sphinxext extension is
"to document drivers and other types of plugins to make the available
sets built into projects easier to discover"(dhellman). For the heat 
project, since there are so many plugins that need documentation, 
we discussed having the initial/parent patch be the enabling 
of the extension alone, then following up with individual patches 
that add the appropriate documentation to each plugin that requires it.
The next patch in the set will be one of the plugins that needs 
documentation, and it will depend on this patch that enables the 
extension.

More details in the ML thread at 
http://lists.openstack.org/pipermail/openstack-dev/2015-August/073338.html

Change-Id: Id05f80c2221c66ed3e2561983654e19bb253c572