Updated openstack/openstack Project: openstack/ironic  d5dc270b6e9bb654174f6f46f3fbd1b4428c561b
Clean up cleaning error handling on heartbeats

If a node is in cleaning and an error occurs during the heartbeat
handler, the cleaning_error_handler is not always called (e.g. if
Ironic cannot reach the agent). This means the network
is not torn down properly and the clean_step is not reset. Fix the
error message that is set in last_error when an error occurs during
cleaning.

Change-Id: I22af61b641a49b48b36c8e57750d8db8a9b8cc94
Closes-Bug: #1485667