Updated openstack/openstack Project: openstack/neutron  7a335edc9b67f1c1a73027fe01a9ea34d176a94c
Catch errors on 'port not found' while deleting subnet

In some circumstances (like the one triggered by the test_dhcp_ipv6 testcase)
calls to deleting a port and calls to deleting subnets can happen in straight
sequence.

If this happens the execution of the operations can interleave leading
to the subnet deletion to fail because the port has already gone. This patch
ensures a missing port is handled correctly.

The method delete_subnet is ginormous and hence impossible to test at a unit
level without proper refactoring. That can happen with a follow-up patch.

Closes-bug: #1490832

Change-Id: I80c3733c93b2b66c2a1c4bc3bc24272afdd88b1f