Updated openstack/openstack Project: openstack/horizon  20fa23a86d27ca6c0d8d38686eeee64345aed428
Run a subset of tests on Python 3

Fix "tox -e py34" by only running a subset of tests which pass on Python
3.4. It should fix the py34 gate. If the gate pass, we can make it
voting to avoid Python 3 regressions.

Then we can add more tests to this list when fixing more Python 3
issues. The final goal is to remove the whitelist when all tests will
pass on Python 3.

Partial-Implements: blueprint porting-python3
Change-Id: I3e56789097162f9efc93b889e54b95147ea1f2dc