Updated openstack/openstack Project: openstack/ironic-python-agent  36ce80175ac7616f0a044019604aece91433d06a
Split functional and py27 tests

Moved tests/extensions to tests/unit/extensions, otherwise they
wouldn't have run. Renamed the files in tests/unit/extensions to be
prefixed with 'test_', to match the rest of the test files. Did a
general clean up of tox.ini so we have standard whitespace.

Change-Id: I5615bfbf2d6ff4a4bc3e66e71c2472b0d1e3a2d7