Updated openstack/openstack Project: openstack/barbican  42c74191714ca818b3be095bea20647ebac27716
Update API Reference Docs with Project Quota Information

Add a new page to the API Reference detailing the Quota and Project
Quota APIs.  Also update the pages for secrets, containers, and order
to add a the new status code that is returned when a quota limit
has been enforced.

Change-Id: I05bb3cbd7d319dd2b2a64cba7b88aba72d4b1914
Partially-implements: blueprint quota-support-on-barbican-resources