Updated openstack/openstack Project: openstack/python-openstackclient  9210cac86a2d5d3df9989f40b1fcc605e5f78ed8
Ignore flavor and image find errors on server show

If there is an error finding an image or a flavor during
image show, ignore it and just print the id of the flavor
or image.  This code is also used during server create and
server rebuild, but only to display the results.

Change-Id: I5362158ab8ffc3e5a0800904d6ea15420c3a8627
Closes-bug: #1489901