Updated openstack/openstack Project: openstack/nova  b32b146a336bed2b8527b718af9015c4e2686917
[Ironic]Not count available resources of deployed ironic node

Changing node's properties in Ironic after node is deployed will
count as available resources in Nova, as nova's resource_tracker
will override the *_used keys we set in the driver.

This patch changes the caculation of available resources based on
the instance flavor instead of node to avoid the issue.

Change-Id: I1a6e4501122c4781e103fd83a146e0914c7671c9
Closes-Bug: #1301279