Updated openstack/openstack Project: openstack/nova  bbe369c9b13117a0df9e605074d92f0a0d4551ef
Fix RequestSpec.instance_group hydration

A mistake was done by hydrating the members field for the InstanceGroup object when
creating the RequestSpec object because that field is for getting the instances,
not the hosts. In order to provide that information, we need to provide a new
version of the InstanceGroup and give a lazy-load for getting it in case we want to
have it.

Change-Id: I0d892bbe490c9ee727f54077d7398cc2800ce107
Partially-Implements: blueprint request-spec-object