Updated openstack/openstack Project: openstack/cinder  db19c6907cd840e6e0af4d2d15c6946006632a69
Implement manage/unmanage snapshot in Pure drivers

This adds implementations for manage_existing_snapshot,
manage_existing_snapshot_get_size, and unmanage_snapshot for
PureFCDriver and PureISCSIDriver.

Similar to the implementation of the manage/unmanage volume these will
need a volume_ref containing a ‘name’ key and value that matches an
object in Purity. In this case it will need to be the full name of a
snapshot object.

Due to this requiring a rename of the snapshot it will only be available
if using Purity REST API v1.4 or greater. If using a lower version an
exception will be raised.

Change-Id: Icb788e6cc0ee0e13393a12bc6aaccc57e1d491fd
Implements: blueprint pure-manage-snapshot