Updated openstack/openstack Project: openstack/ironic  e1a1024dd5b5f90383e92e6bc1214a0a31f68127
Revert "Do not overwrite the iPXE boot script on every deployment"

I agree with what Ruby pointed out at [1] about this change. Not
replacing the file can have bigger consequences such not getting
overwrited even when ironic-conductor is updated (in case we change the
boot.ipxe in the future).

So we should revert this patch and deal with the pain of having to
change the ipxe_boot_script config file and restart the ironic-conductor
if we want to change a custom version of the script.

[1] https://review.openstack.org/#/c/218290/2/ironic/drivers/modules/pxe.py

This reverts commit d72493deb21db90961c80757fd81d30fb39a43c8.

Change-Id: I787edfad66d3a91c9b106996a0ed5f3d7f624727