Updated openstack/openstack Project: openstack/neutron  fae3d2dceacb935a64536d0f26af431ccd566d7c
Make Neutron service flavor save service_type

While the service_type exists in the resource attributes and as
a database field for a Flavor, the creation dictionary did not
pass the value so the service_type was not being persisted
in the database nor returned.

Enhanced unit test to show problem.  Test fails on old code
to save or return the input service_type.

Change-Id: I4dba287f5972ecebd193d65e7f542dd0a65f055b
Closes-Bug: 1490063