Updated openstack/openstack Project: openstack/neutron  4cbcb2eba4502b7fd65abb7c9af38cf02a38cc96
Deprecate --service option for neutron-db-manage

Now that https://review.openstack.org/198542 has merged we can
deprecate the --service option. From now on instead of
  --service fwaas
we should instead use
  --subproject neutron-fwaas

This puts the *aas subprojects on equal footing with the other
projects in the Neutron Stadium for neutron-db-manage.

In the Liberty release the --service option will be marked as
deprecated. It will be removed in Mitaka.

Related-Bug: #1470625

Change-Id: Iecc678efafd798b62bb83e6e85333c64760f16b5