Updated openstack/openstack Project: openstack/neutron  3ce9473bab1c054aa027ca9f84e8a2586e297958
Add info to debug test_keepalived_respawns gate failure

Current theory is that there's a bug in external_process.active,
it returns True when it shouldn't, then kill -15 on the process
pid fails because the process isn't up. Added ps -p output to
see if the process is up or not.

Change-Id: Ic062be829d5f05a1294f6b2fa54820422871ffcb
Related-Bug: #1490043