Updated openstack/openstack Project: openstack-infra/project-config  eaa87031bb966a89735a0bdab001e2d5476941c6
Experimental multinode job for puppet-openstack-integration

This is the first step needed to run a multinode environment for
puppet-openstack-integration module. For now, this is experimental and
will require some time to flesh out some of nodepool functionality.

Currently run_multinode.sh does not exist, however once this gate has
been added, we'll be able to upload our patch containing it.  This
will give the puppet openstack team a simple multinode job to work
with to better understand nodepool subnodes lifecycle.

Change-Id: Ibac4c9af362d9a6c548b9d0d672f9ace3020f0cc
Signed-off-by: Paul Belanger <pabelanger@redhat.com>