Updated openstack/openstack Project: openstack/barbican  b30fd9cdf4089ce8d3952493ddfd5fe808894f8c
Stanardized Functional Tests

The functional tests included some tests for creating AES keys with
different bit lengths. Most of those bit lengths were not valid. The
tests can cause some secret stores to return failures because the key
lengths were not valid. The tests were also setting the bit length to
different values but not modifying the payload. The payload is now
being created for the appropriate bit length.

One of the tests was passing the algorithm and bit length as part of a
storage operation. The algorithm name was invalid. This can cause some
failures on devices that do not support that algorithm.

The tests that use non-standard algorithms or bit lengths were
decorated with a non-standard-algorithm attribute. This will allow
testers to skip those tests when selecting which ones to run.

Change-Id: Ia425abef9232cd6729a23e87cbf8153f211a49a3