Updated openstack/openstack Project: openstack/horizon  463990c25bdcc979acb4a0d77f834a26fb5667c8
Use security group id not name in the create instance dialog

It's possible to create multiple security groups with the same name,
and then launching an instance will error with "Error: Multiple
security_group matches found for name 'foo', use an ID to be more specific."

The solution is to use the group id instead of group name in the form.

This only works with Neutron, because Nova-Network not only requires the
groups to be passed by name, but also passes them by name internally in
several places.

Change-Id: I83587e15a8c18d64a29fd1083e786a73162b19f9
Partial-bug: #1203413