Updated openstack/openstack Project: openstack/heat  199da1057582bcfc6e94fa6e064f2dc0f2de255f
Set correct tenant id when python-client init

If tenant name with chinese characters, an
'UnicodeEncodeError' exception raised when the tenant
creating a stack with a nova server. To set correct
tenant id instead of tenant name when novaclient init
is the best way to avoid this problem.

Also this patch fix this problem for glanceclient and
cinderclient.

Change-Id: Ifc8ab70c69453344156d25aba4291769bb7fe117
Closes-Bug: #1488881