Updated openstack/openstack Project: openstack-infra/system-config  aac99f58f897237f1e33bae16356206d1261d681
Switch to centos7 git01 and git02 hosts

Continue the great switch to centos7 for git backends and replace git01
and 02. Since these are the last two backends to be swapped out switch
the balance method back to leastconn as we will go back to having
homogenous git backends and can get away with leastconn balancing.

Change-Id: Ib3180587892bb46b3dc9d6a7dea28a28da85c3b3