Updated openstack/openstack Project: openstack/trove  8bec82524db282e38dddc5d7e91ab017fcf80494
Implements Datastore Registration API

Adds the mgmt APIs for

1. Add a new datastore-version.
2. Reset the attributes of an existing datastore-version.
3. List datastore-versions.
4. Get details of a datastore-version.
5. Delete existing datastore-version.

In order to simplify the process of registering an image for a new
datastore/datastore-version the datastore would be managed behind the
scenes. This means that when creating a datastore-version, a datastore
will be automatically created, if it does not already exist.

DocImpact
Implements: blueprint datastore-registration-api
Depends-On: I915baa4c5c8bbe3f0483bcaeaa5ce95044d35bf7
Change-Id: If79a03962abc41033dfe6e6ec1e335b243bb74f7